xfs: make AIL target updates and compares 32bit safe.
Patchwork-id: 39913
O-Subject: [REHL6.2 PATCH 340/361] xfs: make AIL target updates and compares 32bit safe.
Bugzilla: 695827
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>

RH BZ: 695827 - XFS update to 2.6.39 code base for delayed logging
Upstream Commit: fd5670f22fce247754243cf2ed41941e5762d990

Upstream commit message:

The recent conversion of the xfsaild functionality to a work queue
introduced a hard-to-hit log space grant hang. One of the problems
noticed was that updates of the push target are not 32 bit safe as
the target is a 64 bit value.

We cannot copy a 64 bit LSN without the possibility of corrupting
the result when racing with another updating thread. We have
function to do this update safely without needing to care about
32/64 bit issues - xfs_trans_ail_copy_lsn() - so use that when
updating the AIL push target.

Also move the reading of the target in the push work inside the AIL
lock, and use XFS_LSN_CMP() for the unlocked comparison during work
termination to close read holes as well.

Signed-off-by: Dave Chinner <david@fromorbit.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Alex Elder <aelder@sgi.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>