xfs: reset inode per-lifetime state when recycling it
Patchwork-id: 39907
O-Subject: [REHL6.2 PATCH 358/361] xfs: reset inode per-lifetime state when recycling it
Bugzilla: 695827
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>

RH BZ: 695827 - XFS update to 2.6.39 code base for delayed logging
Upstream Commit: 778e24bb6dd8682318bb496d4bfdc32b501a6420

Upstream commit message:

XFS inodes has several per-lifetime state fields that determine the
behaviour of the inode. These state fields are not all reset when an
inode is reused from the reclaimable state.

This can lead to unexpected behaviour of the new inode such as
speculative preallocation not being truncated away in the expected
manner for local files until the inode is subsequently truncated,
freed or cycles out of the cache. It can also lead to an inode being
considered to be a filestream inode or having been truncated when
that is not the case.

Rework the reinitialisation of the inode when it is recycled to
ensure that it is pristine before it is reused. While there, also
fix the resetting of state flags in the recycling error paths so the
inode does not become unreclaimable.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Alex Elder <aelder@sgi.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>