reject a re-add request that cannot be honoured
Patchwork-id: 39934
O-Subject: [Patch RHEL6.2 04/30] md: reject a re-add request that cannot be honoured.
Bugzilla: 729696
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: NeilBrown <neilb@suse.de>

The 'add_new_disk' ioctl can be used to add a device either as a
spare, or as an active disk that just needs to be resynced based on
write-intent-bitmap information (re-add)

Currently if a re-add is requested but fails we add as a spare
instead.  This makes it impossible for user-space to check for
failure.

So change to require that a re-add attempt will either succeed or
completely fail.  User-space can then decide what to do next.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit bedd86b7773fd97f0d708cc0c371c8963ba7ba9a)

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>