raid1: improve handling of pages allocated for write-behind
Patchwork-id: 39946
O-Subject: [Patch RHEL6.2 13/30] md/raid1: improve handling of pages allocated for write-behind.
Bugzilla: 729696
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: NeilBrown <neilb@suse.de>

The current handling and freeing of these pages is a bit fragile.
We only keep the list of allocated pages in each bio, so we need to
still have a valid bio when freeing the pages, which is a bit clumsy.

So simply store the allocated page list in the r1_bio so it can easily
be found and freed when we are finished with the r1_bio.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit af6d7b760c7547c1a410a333bdb24daed24e1043)

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>