Add missing SPC-4 CDB and MAINTENANCE_[IN, OUT] service action definitions
Patchwork-id: 39970
O-Subject: [RHEL6.2 PATCH 08/13] [SCSI] Add missing SPC-4 CDB and
	MAINTENANCE_[IN, OUT] service action definitions
Bugzilla: 638007
RH-Acked-by: Mike Christie <mchristi@redhat.com>

This patch adds a handful of missing CDBs defs that are used by TCM
persistent reservation logic in the SPC-4 defined CDB exclusion table for
registrations and reservations.

This includes a number of missing MI_* and MO_* prefixed service actions defs
for MAINTENANCE_IN and MAINTENANCE_OUT that are mentioned wrt to persistent
registration and reservation status for the SCSI Initiator Port.

Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>
Signed-off-by: James Bottomley <James.Bottomley@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>