UV: Avoid KABI breakage
Patchwork-id: 39992
O-Subject: Re: [RHEL6.2 PATCH 03/12] x86, UV: Avoid KABI breakage
Bugzilla: 696269
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

bz696269 UV2 Hub Support
second post

This is a somewhat cleaned-up version of the kabi-hack.
We agree its an ugly hack, but necessary for handling
the change from UV1 to UV2.

George
The UV2 patch that is being sent upstream breaks kernel KABIs.

This patch adds code to the UV2 infrastructure so that UV2
support will not break the KABI. The changes are ugly but not intrusive.

With these changes, out-of-tree kernel modules built with UV2 support
will continue to load and on systems running GA kernels. They obviously
will NOT support UV2 but UV1 will work correctly.

This patch goes to distros ONLY. Not for upstream.

Signed-off-by: Jack Steiner <steiner@sgi.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>