IB/srp: rework mapping engine to use multiple FMR entries
Patchwork-id: 40081
O-Subject: [Patch RHEL6.2 83/88] IB/srp: rework mapping engine to use multiple
	FMR entries
Bugzilla: 688944 690152 698275 698279 698283 700277 707092 710159
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: David Dillow <dillowda@ornl.gov>

Instead of forcing all of the S/G entries to fit in one FMR, and falling
back to indirect descriptors if that fails, allow the use of as many
FMRs as needed to map the request. This lays the groundwork for allowing
indirect descriptor tables that are larger than can fit in the command
IU, but should marginally improve performance now by reducing the number
of indirect descriptors needed.

We increase the minimum page size for the FMR pool to 4K, as larger
pages help increase the coverage of each FMR, and it is rare that the
kernel would send down a request with scattered 512 byte fragments.

This patch also move some of the target initialization code afte the
parsing of options, to keep it together with the new code that needs to
allocate memory based on the options given.

Signed-off-by: David Dillow <dillowda@ornl.gov>
(cherry picked from commit 8f26c9ff9cd0317ad867bce972f69e0c6c2cbe3c)

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>