IB/srp: add support for indirect tables that don't fit in SRP_CMD
Patchwork-id: 40057
O-Subject: [Patch RHEL6.2 84/88] IB/srp: add support for indirect tables that
	don't fit in SRP_CMD
Bugzilla: 688944 690152 698275 698279 698283 700277 707092 710159
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: David Dillow <dillowda@ornl.gov>

This allows us to guarantee the ability to submit up to 8 MB requests
based on the current value of SCSI_MAX_SG_CHAIN_SEGMENTS. While FMR will
usually condense the requests into 8 SG entries, it is imperative that
the target support external tables in case the FMR mapping fails or is
not supported.

We add a safety valve to allow targets without the needed support to
reap the benefits of the large tables, but fail in a manner that lets
the user know that the data didn't make it to the device. The user must
add "allow_ext_sg=1" to the target parameters to indicate that the
target has the needed support.

If indirect_sg_entries is not specified in the modules options, then
the sg_tablesize for the target will default to cmd_sg_entries unless
overridden by the target options.

Signed-off-by: David Dillow <dillowda@ornl.gov>
(cherry picked from commit c07d424d6118d528ef71b22b7424bfc359c307a5)

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>