spec: update ldconfig-kernel.conf for all kernels
Patchwork-id: 40085
O-Subject: [RHEL6 PATCH] spec: update ldconfig-kernel.conf for all kernels
Bugzilla: 708350
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

Bugzilla: 708350

Brew build: https://brewweb.devel.redhat.com/taskinfo?taskID=3526438

The "nosegneg" hwcap is assigned to bit 1 in the 32-bit vDSO:

    $ git grep -i no[ns]eg[ns]eg
    vdso/vdso32/note.S: .byte VDSO_NOTE_NONEGSEG_BIT; .asciz "nosegneg" /* bit, name */
    xen/vdso.h:#define VDSO_NOTE_NONEGSEG_BIT 1

However, for historical reasons the spec is associating it to bit 0.
This may cause a substantial performance hit when running 32-bit RHEL6
Xen PV guests on 32-bit hosts (50-100x on artificial testcases).

Simply fixing the newly-installed .conf file doesn't work because hwcaps
must not overlap.  This message is then shown after two conflicting
kernels are installed, every time ldconfig is run:

        ldconfig: /etc/ld.so.conf.d/kernel-2.6.32-170.el6.x86_64.conf:6: hwcap index 0 already defined as nosegneg

I think this should not really be an issue, but good luck convincing
the glibc maintainer.  So, let's instead update all .conf files.
This is incorrect if a non-pvops, RH-packaged Xen kernel is installed
on the machine, but the incorrectness is not fatal and the only such
kernels in existence are RHEL5 vintage.  Which makes this patch
Good Enough.

Please review and ack.

Paolo

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>