Add support of Synaptics Clickpad device
Patchwork-id: 40100
O-Subject: [RHEL6 BZ#464653 02/15] Input: Add support of Synaptics Clickpad device
Bugzilla: 464653
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Upstream changeset: 5f57d67

Author: Takashi Iwai <tiwai@suse.de>
Date:   Mon Apr 19 10:37:21 2010 -0700

The new type of touchpads can be detected via a new query command
0x0c. The clickpad flags are in cap[0]:4 and cap[1]:0 bits.

When the device is detected, the driver now reports only the left
button as the supported buttons so that X11 driver can detect that
the device is Clickpad. A Clickpad device gives the button events
only as the middle button. The kernel driver morphs to the left
button. The real handling of Clickpad is done rather in X driver
side.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>