netxen-firmware: rename firmware files
Patchwork-id: 40125
O-Subject: [PATCH RHEL6.2] netxen-firmware: rename firmware files
Bugzilla: 693743
RH-Acked-by: Don Zickus <dzickus@redhat.com>

When storing firmware in /lib/firmware we need to use file names that
contain version strings to ensure kernel upgrades and downgrades still
work properly.  This file is renamed as compared to upstream and each
time we update it, we will need to rename it with the new version
number.

This will complete the request in RHBZ 693743.