autofs: fix some further kabi inconsistencies
Patchwork-id: 40195
O-Subject: [RHEL-6.2 PATCH] kabi - fix some further kabi inconsistencies (aka. patch 37/36)
Bugzilla: 704941
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Jeff Moyer <jmoyer@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=704941

The main kABI problem is the introduction of two dentry operation callback
functions. To quiet the kABI check they need to be excluded from the ksyms
generation. These functions have been added to the end of the dentry
operations structure nad the VFS attempts to access them only if the file
system sets ->d_flags to tell it that it wants to use automount functions.
Which means that structure offsets are not changed and out of bounds accesses
will not occur for modules unaware of these changes. So the exclusion of the
symbols from the ksyms generation is valid.

Another change is to not rename the now unused dentry field, d_mounted, since
that may affect the kABI check. This field should not be used directly, it
should be accessed via the d_mountpoint() helper. Also the type of the
d_mountpoint() function was changed from int to bool, which is changed back.

Finally, an unused declaration was removed from include/linux/mount.h and
although it shouldn't cause a kABI problem there is no specific reason to
remove it so it has been added back.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Acked-by: Ian Kent <ikent@redhat.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>