kernel: NSS creation with initrd fails
Patchwork-id: 40261
O-Subject: [RHEL6.2 PATCH 3/3] [s390x]kernel: NSS creation with initrd fails
Bugzilla: 730780

Description
-----------
When IPLing from a block device in order to create an NSS which
includes an initrd this might fail with a message that indicates
that different memory segment overlap.

By default zipl makes sure that the initrd will be loaded right
behind the kernel image, if an IPL from a block device is issued.
If a user wants to create an NSS which includes an initrd this can
be a problem. When creating an NSS different types of page ranges
must be in different segments. Since this is not necessarily given
if the initrd is placed right behind the kernel, the request might
fail.

Move the initrd to a higher address so there are no overlapping
page ranges with different types within any segment.

Bugzilla
--------
BZ 730780
https://bugzilla.redhat.com/show_bug.cgi?id=730780

Upstream status of the patch
----------------------------
The patch is queued for inclusion into a -rc release and is already
available in the maintainers tree at:
http://148.100.100.18/cgi-bin/gitweb.cgi?p=linux-2.6.git;a=shortlog;h=refs/heads/fixes

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>