nfsd4: return nfserr_symlink on v4 OPEN of non-regular file
Patchwork-id: 40316
O-Subject: [RHEL6.2 PATCH] nfsd4: return nfserr_symlink on v4 OPEN of
	non-regular file
Bugzilla: 697659
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Without this, an attempt to open a device special file without first
stat'ing it will fail.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>

Upstream commit: aadab6c6f4da38d639394de740602f146c88da0c
	(committed only to git://linux-nfs.org/~bfields for-3.2 for now.
	Expected to be merged into upstream with that same commit id in
	the 3.2 merge window.)
Bugzilla: 697659

This was a case that was unfortunately left unspecified in the rfc's.
This error return is what works for existing linux nfs clients, so
it appears to be the de facto choice.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>